% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_end}
\alias{adjust_end}
\title{Adjust end generic dispatch}
\usage{
adjust_end(x, adj_by)
}
\arguments{
\item{x}{object to adjust on}

\item{adj_by}{time in seconds to adjust by}
}
\value{
item with new end time
}
\description{
Adjust end generic dispatch
}
