% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_start}
\alias{adjust_start}
\title{Adjust start generic dispatch}
\usage{
adjust_start(x, adj_by)
}
\arguments{
\item{x}{item to adjust on}

\item{adj_by}{time to adjust by in seconds}
}
\value{
item with new start time
}
\description{
Adjust start generic dispatch
}
