% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{create_monoexpgg}
\alias{create_monoexpgg}
\title{Create data.frame of values of monoexponential curve from coefficients}
\usage{
create_monoexpgg(xo, xf, xbin, Amp, Baseline, TD, tau)
}
\arguments{
\item{xo}{start time}

\item{xf}{end time}

\item{xbin}{time increments}

\item{Amp}{amplitude}

\item{Baseline}{baseline}

\item{TD}{time delay}

\item{tau}{time constant}
}
\value{
data.frame
}
\description{
Create data.frame of values of monoexponential curve from coefficients
}
