% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_file_dne}
\alias{error_file_dne}
\title{Error File Does Not Exist}
\usage{
error_file_dne(file)
}
\arguments{
\item{file}{filepath or file name}
}
\value{
message for rlang::abort
}
\description{
Error File Does Not Exist
}
