% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_file_metadata}
\alias{error_file_metadata}
\title{Error File Metadata Missing}
\usage{
error_file_metadata(type)
}
\arguments{
\item{type}{type of metadata}
}
\value{
message for rlang::abort
}
\description{
Error File Metadata Missing
}
