% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_neg_argument}
\alias{error_neg_argument}
\title{Error Negative Argument}
\usage{
error_neg_argument(arg)
}
\arguments{
\item{arg}{argument}
}
\value{
message for rlang::abort
}
\description{
Error Negative Argument
}
