% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ext}
\alias{extract_ext}
\title{Get the file extension type}
\usage{
extract_ext(file)
}
\arguments{
\item{file}{name of the file}
}
\value{
file extension type
\itemize{
\item csv
\item xlsx
\item txt
}
}
\description{
Get the file extension type
}
