% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{facet_bottomup}
\alias{facet_bottomup}
\title{ggplot2 facet from the bottom to the top}
\usage{
facet_bottomup(
  facets,
  nrow,
  ncol,
  strip.position = "top",
  labeller = "label_value"
)
}
\arguments{
\item{facets}{facets}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{strip.position}{where to put the facet strip; default to 'top'}

\item{labeller}{labeller for facets; defaults to 'label_value'}
}
\value{
ggplot2 ggproto object
}
\description{
ggplot2 facet from the bottom to the top
}
