% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{geom_monoexp}
\alias{geom_monoexp}
\title{ggplot2 geom for a monoexponential curve}
\usage{
geom_monoexp(
  mapping = NULL,
  data = NULL,
  stat = "monoexp",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xo = 0,
  xf = 120,
  xbin = 1,
  ...
)
}
\arguments{
\item{mapping}{ggplot2 mapping}

\item{data}{ggplot2 data}

\item{stat}{ggplot2 stat name}

\item{position}{ggplot2 position}

\item{na.rm}{remove nas}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit from parent}

\item{xo}{time start}

\item{xf}{time end}

\item{xbin}{time increments}

\item{...}{other parameters}
}
\value{
ggplot2 stuff
}
\description{
ggplot2 geom for a monoexponential curve
}
