% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{geom_monoexptau}
\alias{geom_monoexptau}
\title{ggplot2 geom for monoexpoential curve with tau points}
\usage{
geom_monoexptau(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{ggplot2 mapping}

\item{data}{ggplot2 data}

\item{stat}{ggplot2 stat}

\item{position}{ggplot2 position}

\item{...}{other parameters to pass}

\item{na.rm}{remove na}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit aes}
}
\value{
ggplot2 layer
}
\description{
ggplot2 geom for monoexpoential curve with tau points
}
