% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{geom_tau}
\alias{geom_tau}
\title{ggplot2 custom geom for tau as a point on a line}
\usage{
geom_tau(
  mapping = NULL,
  data = NULL,
  stat = "tau",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{ggplot2 mapping}

\item{data}{ggplot2 mapping}

\item{stat}{ggplot2 stat}

\item{position}{ggplot2 position}

\item{...}{other parameters}

\item{na.rm}{remove nas}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit from parent}
}
\value{
ggplot2 stuff
}
\description{
ggplot2 custom geom for tau as a point on a line
}
