% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ch_names_by_type}
\alias{get_ch_names_by_type}
\title{Get a list of channel names by channel type}
\usage{
get_ch_names_by_type(info)
}
\arguments{
\item{info}{info object}
}
\value{
list with names "ch_type" and value character vector of
respective "ch_names
}
\description{
Get a list of channel names by channel type
}
