% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_ch_type}
\alias{get_ch_type}
\title{Get channel types from column names}
\usage{
get_ch_type(longname)
}
\arguments{
\item{longname}{}
}
\value{
ch_type
}
\description{
Get channel types from column names
}
