% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_id_from_filename}
\alias{get_id_from_filename}
\title{Get participant ID from the file name}
\usage{
get_id_from_filename(file)
}
\arguments{
\item{file}{string}
}
\value{
string
}
\description{
Get participant ID from the file name
}
