% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_labels}
\alias{get_labels}
\title{Make usable short labels for columns from the long names given by oxysoft}
\usage{
get_labels(longname)
}
\arguments{
\item{longname}{character string}
}
\value{
shorter character string
}
\description{
Make usable short labels for columns from the long names given by oxysoft
}
