% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from scanned list}
\usage{
get_metadata(s, type)
}
\arguments{
\item{s}{scanned list of data as characters}

\item{type}{type of metadata to look for}
}
\value{
value
}
\description{
Get metadata from scanned list
}
