% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_class.R
\name{get_transitions}
\alias{get_transitions}
\title{Get locations of transitions to analyze}
\usage{
get_transitions(protocol)
}
\arguments{
\item{order}{vector of the order of each type of exercise in the protocol}
}
\value{
list containing start and end times
}
\description{
Get locations of transitions to analyze
}
