% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{gg_monoexp}
\alias{gg_monoexp}
\title{Monoexponential function from coefficients}
\usage{
gg_monoexp(Amp, Baseline, TD, tau)
}
\arguments{
\item{Amp}{amplitude}

\item{Baseline}{baseline}

\item{TD}{time delay}

\item{tau}{time constant}
}
\value{
function
}
\description{
Monoexponential function from coefficients
}
