% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_info.R
\name{info}
\alias{info}
\title{Info Class}
\usage{
info(
  chs = data.frame(),
  device_info = NULL,
  subj_info = NULL,
  meas_date = NULL,
  meas_start = NULL,
  sfreq = NULL,
  samps = "samp_num",
  bads = NULL
)
}
\arguments{
\item{chs}{data frame of channels in measurement where rows are channels
and columns are ch_name, ch_type, source_num, det_num, wavelength}

\item{device_info}{list containing device information
device type, model, serial, site}

\item{subj_info}{list containing subject information}

\item{meas_date}{datetime of the measurement}

\item{meas_start}{time to crop off beginning of file}

\item{sfreq}{sampling frequency in Hz}

\item{samps}{name of the column containing sample numbers. Defaults to
"samp_num" which is the default name if you use nirsex::read_nirs,}

\item{bads}{list of channels that have bad signal}
}
\value{
info

Contains all metadata associated with the recording.
}
\description{
Info Class
}
