% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexp.R
\name{monoexp}
\alias{monoexp}
\title{Create data.frame of monoexponential y values at time x based on coefficients}
\usage{
monoexp(Amp, Baseline, TD, tau, xlen)
}
\arguments{
\item{Amp}{amplitude}

\item{Baseline}{baseline y}

\item{TD}{time delay}

\item{tau}{time constant}

\item{xlen}{number of output x values}
}
\value{
data.frame
}
\description{
Create data.frame of monoexponential y values at time x based on coefficients
}
