% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_class.R
\name{new_protocol}
\alias{new_protocol}
\title{Protocol Class Constructor}
\usage{
new_protocol(p = data.frame())
}
\arguments{
\item{p}{data frame}
}
\value{
data frame of class protocol
}
\description{
Protocol Class Constructor
}
