% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rawdata.R
\name{new_rawdata}
\alias{new_rawdata}
\title{Rawdata Class Constructor}
\usage{
new_rawdata(data = data.frame(), info = info())
}
\arguments{
\item{data}{data frame}

\item{info}{measurement metadata}
}
\value{
data frame of raw data
}
\description{
Rawdata Class Constructor
}
