% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampletime.R
\name{new_sampletime}
\alias{new_sampletime}
\title{nirsex_sampletime record vector constructor}
\usage{
new_sampletime(
  sampnum = integer(),
  orig_min = integer(),
  orig_sec = integer(),
  orig_ms = integer(),
  zeroed_min = integer(),
  zeroed_sec = integer(),
  zeroed_ms = integer()
)
}
\arguments{
\item{sampnum}{sample number}

\item{orig_min}{minute of sample}

\item{orig_sec}{second of sample}

\item{orig_ms}{millisecond of sample}

\item{zeroed_min}{minute of sample after zeroed sample}

\item{zeroed_sec}{second of sample after zeroed sample}

\item{zeroed_ms}{millisecond of sample after zeroed sample}
}
\value{
nirsex_sampletime
}
\description{
list representation of the time since starting the measurement.
}
\details{
displays as
if sample zero is the first sample, then zeroed_min, zeroed_sec
and zeroed_ms will match orig_min, orig_sec, and orig_ms respectively

if sample zero is not the first sample, then zeroed_min, zeroed_sec,
and zeroed_ms will refer to the time since sample zero
and orig_min, orig_sec, and orig_ms will refer to the original time
}
