% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nirsex_example}
\alias{nirsex_example}
\title{Access nirsex example files}
\usage{
nirsex_example(path = NULL)
}
\arguments{
\item{path}{filename}
}
\value{
path to file
}
\description{
Access nirsex example files
}
\examples{
read_nirs_example()
read_nirs_example("opticaldensity_raw.txt")
}
