% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_all_raw_transitions}
\alias{plot_all_raw_transitions}
\title{Display a plot of all raw transition data}
\usage{
plot_all_raw_transitions(l, cols)
}
\arguments{
\item{l}{list of timealigned transition data}

\item{cols}{cols to facet by}
}
\value{
cowplot
}
\description{
Display a plot of all raw transition data
}
