% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_all_results}
\alias{plot_all_results}
\title{plot results of monoexponential regression}
\usage{
plot_all_results(results, cols)
}
\arguments{
\item{results}{list of results}
}
\value{
cowplot
}
\description{
plot results of monoexponential regression
}
