% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_averageddata}
\alias{plot_averageddata}
\title{plot averaged data}
\usage{
plot_averageddata(dt, timecol = "Time", cols)
}
\arguments{
\item{dt}{averaged across signals}

\item{timecol}{column that contains the time}

\item{cols}{cols to facet}
}
\value{
ggplot object
}
\description{
plot averaged data
}
