% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_one_result}
\alias{plot_one_result}
\title{plot result from a single monoexponential regression}
\usage{
plot_one_result(resultslist, signal, timecol = "Time")
}
\arguments{
\item{resultslist}{list of regression output}

\item{signal}{nirs signal type}

\item{timecol}{name of time column}
}
\value{
ggplot2
}
\description{
plot result from a single monoexponential regression
}
