% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_rawtransitions}
\alias{plot_rawtransitions}
\title{plot raw data for a transition type}
\usage{
plot_rawtransitions(timealigneddt, timecol = "Time", cols, name)
}
\arguments{
\item{timealigneddt}{data table that has been time aligned
with columns timecol, group, and variable cols
where group is a character describing which transition
the data is in}

\item{timecol}{character of column name containing times}

\item{cols}{variable columns to facet}
}
\value{
ggplot
}
\description{
plot raw data for a transition type
}
