% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{plot_results}
\alias{plot_results}
\title{Create a plot for the results of one transition type}
\usage{
plot_results(resultslist, cols, timecol = "Time")
}
\arguments{
\item{resultslist}{list of results for the transition type}

\item{cols}{variables to facet by}

\item{timecol}{column of time data}
}
\value{
ggplot
}
\description{
Create a plot for the results of one transition type
}
