% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_class.R
\name{print.nirsex_protocol}
\alias{print.nirsex_protocol}
\title{Print Generic for protocol class}
\usage{
\method{print}{nirsex_protocol}(p)
}
\arguments{
\item{p}{protocol}
}
\value{
print to console
}
\description{
Print Generic for protocol class
}
