% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_class.R
\name{protocol}
\alias{protocol}
\title{Protocol Class Helper}
\usage{
protocol(descriptions = character(), duration = double(), order = NULL)
}
\arguments{
\item{descriptions}{character vector of section descriptions}

\item{duration}{double vector of section durations}

\item{order}{optional integer vector}
}
\value{
protocol
}
\description{
Protocol Class Helper
}
