% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rawdata.R
\name{rawdata}
\alias{rawdata}
\title{Rawdata Class Helper}
\usage{
rawdata(
  data = data.frame(),
  chs = data.frame(),
  device_info = NULL,
  subj_info = NULL,
  meas_date = NULL,
  meas_start = NULL,
  sfreq = NULL,
  samps = "samp_num",
  bads = NULL
)
}
\arguments{
\item{data}{data frame of raw data}

\item{info}{measurement metadata}
}
\value{
rawdata
}
\description{
Rawdata Class Helper
}
