% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nirs.R
\name{read_nirs}
\alias{read_nirs}
\title{NIRS Data Input}
\usage{
read_nirs(file, ...)
}
\arguments{
\item{file}{file to read}
}
\value{
data.frame
}
\description{
Reads a file of NIRS data and creates a data frame from it.
}
