% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nirs.R
\name{read_nirs.txt}
\alias{read_nirs.txt}
\title{Read a .txt file of NIRS data}
\usage{
\method{read_nirs}{txt}(file, sep = "\\t", colnames = "short")
}
\arguments{
\item{file}{filepath}

\item{sep}{column delimiter}
}
\value{
data frame
}
\description{
Read a .txt file of NIRS data
}
