% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{stat_monoexp}
\alias{stat_monoexp}
\title{ggplot2 stat for monoexponential curves}
\usage{
stat_monoexp(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xo = 0,
  xf = 120,
  xbin = 1,
  ...
)
}
\arguments{
\item{mapping}{ggplot2 mapping}

\item{data}{ggplot2 data}

\item{geom}{ggplot2 geom}

\item{position}{ggplot2 position}

\item{na.rm}{remove na}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit parent aes}

\item{xo}{xstart}

\item{xf}{xend}

\item{xbin}{x increment by}

\item{...}{other parameters}
}
\value{
ggplot2 layer
}
\description{
ggplot2 stat for monoexponential curves
}
