% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nirsex.R
\name{stat_tau}
\alias{stat_tau}
\title{ggplot2 stat tau points}
\usage{
stat_tau(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{ggplot2 mapping}

\item{data}{ggplot2 data}

\item{geom}{ggplot2 geom}

\item{position}{ggplot2 position}

\item{...}{other parameters}

\item{na.rm}{remove na}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit aes}
}
\value{
ggplot2 layer
}
\description{
ggplot2 stat tau points
}
