% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rawdata.R
\name{validate_rawdata}
\alias{validate_rawdata}
\title{Rawdata Class Validator}
\usage{
validate_rawdata(data, info)
}
\arguments{
\item{data}{data frame}

\item{info}{measurement metadata}
}
\value{
data
}
\description{
Rawdata Class Validator
}
